+++ packages/react-behave/src/components/ClickOutside/ClickOutside.js
@@ -10,12 +10,12 @@ import EventListener from 'react-event-listener'
  * ## Usage
  *
  * ```jsx
- * import React, { Component } from 'react';
- * import { ClickOutside } from 'react-behave';
+ * import React, { Component } from 'react'
+ * import { ClickOutside } from 'react-behave'
  *
  * class App extends Component {
  *   handleClickOutside(event) {
- *     console.log('click', event);
+ *     console.log('click', event)
  *   }
  *
  *   render() {
@@ -23,7 +23,7 @@ import EventListener from 'react-event-listener'
  *       <ClickOutside onClickOutside={this.handleClickOutside}>
  *         {ref => <p ref={ref}>Don't click on me.</p>}
  *       </ClickOutside>
- *     );
+ *     )
  *   }
  * }
  * ```