+++ packages/react-behave/src/components/Dropdown/Dropdown.js
@@ -4,18 +4,19 @@ import { Manager, Popper, Reference, placements } from 'react-popper'
 import ClickOutside from '../ClickOutside'
 import MergeRefs from '../MergeRefs'
 import minWidthModifier from './minWidthModifier'
+import getOtherProps from '../../core/getOtherProps'
 
-function renderContent(onClickOutside, popperProps, popperRef, renderDropDown) {
+function renderContent(renderDropdown, onClickOutside, props) {
   if (!onClickOutside) {
     // Don't add the `ClickOutside` component if nobody is listening.
-    return renderDropDown(popperRef, popperProps)
+    return renderDropdown(props)
   }
 
   return (
     <ClickOutside onClickOutside={onClickOutside}>
       {clickOutsideRef => (
-        <MergeRefs refs={[clickOutsideRef, popperRef]}>
-          {ref => renderDropDown(ref, popperProps)}
+        <MergeRefs refs={[clickOutsideRef, props.ref]}>
+          {ref => renderDropdown({ ...props, ref })}
         </MergeRefs>
       )}
     </ClickOutside>
@@ -27,40 +28,43 @@ function renderContent(onClickOutside, popperProps, popperRef, renderDropDown) {
  * [react-popper]: https://github.com/FezVrasta/react-popper
  * [popper-props]: https://github.com/FezVrasta/react-popper#children
  * [popper-placements]: https://popper.js.org/popper-documentation.html#Popper.placements
+ * [popper-modifiers]: https://popper.js.org/popper-documentation.html#modifiers
  *
  * Render a dropdown around a component.
  *
  * ## Usage
  *
  * ```jsx
- * import React, { Component } from 'react';
- * import { Dropdown } from 'react-behave';
+ * import React, { Component } from 'react'
+ * import { Dropdown } from 'react-behave'
  *
  * class App extends Component {
  *   state = {
  *     open: false,
- *   };
+ *   }
  *
  *   handleClickOutside = () => {
- *     this.setState({ open: false });
- *   };
+ *     this.setState({ open: false })
+ *   }
  *
  *   toggleDropDown = () => {
- *     this.setState(({ open }) => ({ open: !open }));
- *   };
+ *     this.setState(({ open }) => ({ open: !open }))
+ *   }
  *
  *   render() {
  *     return (
  *       <Dropdown
  *         onClickOutside={this.handleClickOutside}
  *         open={this.state.open}
- *         renderDropDown={(ref, { style }) => (
- *           <ul ref={ref} style={style}>
- *             <li>Item 1</li>
- *             <li>Item 2</li>
- *             <li>Item 3</li>
- *           </ul>
- *         )}
+ *         renderDropdown={({ open, ref, style }) =>
+ *           open && (
+ *             <ul ref={ref} style={style}>
+ *               <li>Item 1</li>
+ *               <li>Item 2</li>
+ *               <li>Item 3</li>
+ *             </ul>
+ *           )
+ *         }
  *       >
  *         {ref => (
  *           <button ref={ref} onClick={this.toggleDropDown}>
@@ -68,33 +72,28 @@ function renderContent(onClickOutside, popperProps, popperRef, renderDropDown) {
  *           </button>
  *         )}
  *       </Dropdown>
- *     );
+ *     )
  *   }
  * }
  * ```
  */
-function Dropdown({
-  children,
-  onClickOutside,
-  open,
-  placement,
-  renderDropDown,
-}) {
+function Dropdown(props) {
+  const otherProps = getOtherProps(Dropdown, props)
   return (
     <Manager>
-      <Reference>{({ ref }) => children(ref)}</Reference>
-      {open && (
-        <Popper placement={placement} modifiers={{ minWidthModifier }}>
-          {({ ref: popperRef, ...popperProps }) =>
-            renderContent(
-              onClickOutside,
-              popperProps,
-              popperRef,
-              renderDropDown,
-            )
-          }
-        </Popper>
-      )}
+      <Reference>{({ ref }) => props.children(ref)}</Reference>
+      <Popper
+        {...otherProps}
+        placement={props.placement}
+        modifiers={{ minWidthModifier, ...props.modifiers }}
+      >
+        {popperProps =>
+          renderContent(props.renderDropdown, props.onClickOutside, {
+            ...popperProps,
+            open: props.open,
+          })
+        }
+      </Popper>
     </Manager>
   )
 }
@@ -107,6 +106,16 @@ Dropdown.propTypes = {
    * `ref` must be passed to the component in order to position correctly the dropdown.
    */
   children: PropTypes.func.isRequired,
+
+  /**
+   * Modifiers used to alter the behavior of your poppers.
+   *
+   * In addition to the default modifiers, `minWidthModifier` is added.
+   * It ensures for top or bottom placements that the width of the dropdown is greater or equal to the width of the reference.
+   *
+   * See [PopperJS's modifiers][popper-modifiers].
+   */
+  modifiers: PropTypes.object,
 
   /**
    * _Parameters_: `event: MouseEvent`
@@ -135,10 +144,11 @@ Dropdown.propTypes = {
    * `popperProps` is the object containing styles for positioning that need to be applied on the dropdown component.
    * This object is the same as the [one provided by react-popper][popper-props] without the `ref`.
    */
-  renderDropDown: PropTypes.func.isRequired,
+  renderDropdown: PropTypes.func.isRequired,
 }
 
 Dropdown.defaultProps = {
+  modifiers: {},
   open: false,
   placement: 'bottom-start',
 }