+++ packages/react-behave/src/components/Dropdown/Dropdown.stories.js
@@ -10,28 +10,36 @@ stories.add('Render functions', () => (
   <Dropdown
     onClickOutside={action('onClickOutside')}
     open={boolean('Open', true)}
-    renderDropDown={(ref, { style }) => (
-      <ul ref={ref} style={style}>
-        <li>Item 1</li>
-        <li>Item 2</li>
-        <li>Item 3</li>
-      </ul>
-    )}
+    renderDropdown={({ open, ref, style }) =>
+      open && (
+        <ul ref={ref} style={style}>
+          <li>Item 1</li>
+          <li>Item 2</li>
+          <li>Item 3</li>
+        </ul>
+      )
+    }
   >
-    {ref => <button ref={ref}>Click me</button>}
+    {ref => (
+      <button ref={ref} style={{ width: 300 }}>
+        Click me
+      </button>
+    )}
   </Dropdown>
 ))
 
 stories.add('Without onClickOutside', () => (
   <Dropdown
     open={boolean('Open', true)}
-    renderDropDown={(ref, { style }) => (
-      <ul ref={ref} style={style}>
-        <li>Item 1</li>
-        <li>Item 2</li>
-        <li>Item 3</li>
-      </ul>
-    )}
+    renderDropdown={({ open, ref, style }) =>
+      open && (
+        <ul ref={ref} style={style}>
+          <li>Item 1</li>
+          <li>Item 2</li>
+          <li>Item 3</li>
+        </ul>
+      )
+    }
   >
     {ref => <button ref={ref}>Click me</button>}
   </Dropdown>