+++ packages/react-behave/src/components/Dropdown/Dropdown.test.js
@@ -4,50 +4,47 @@ import ClickOutside from '../ClickOutside'
 import { mount } from 'enzyme'
 
 describe('<Dropdown />', () => {
-  it('passes down the ref to `render`', () => {
-    const render = jest.fn()
-    mount(<Dropdown renderDropDown={jest.fn()}>{render}</Dropdown>)
+  it('passes down the refs', () => {
+    const render = jest.fn(() => null)
+    const renderDropdown = jest.fn(() => null)
+    mount(<Dropdown renderDropdown={renderDropdown}>{render}</Dropdown>)
     expect(render).toHaveBeenCalledWith(expect.anything())
+    expect(renderDropdown).toHaveBeenCalledWith(
+      expect.objectContaining({
+        open: false,
+        ref: expect.anything(),
+        style: expect.any(Object),
+      }),
+    )
   })
 
-  it('passes down the ref and props to `renderDropDown`', () => {
-    const renderDropDown = jest.fn(() => <p />)
+  it('should pass down the open state', () => {
+    const renderDropdown = jest.fn(() => null)
     mount(
-      <Dropdown open={true} renderDropDown={renderDropDown}>
-        {jest.fn()}
+      <Dropdown open={true} renderDropdown={renderDropdown}>
+        {jest.fn(() => null)}
       </Dropdown>,
     )
-    expect(renderDropDown).toHaveBeenCalledWith(
-      expect.anything(),
-      expect.anything(),
+    expect(renderDropdown).toHaveBeenCalledWith(
+      expect.objectContaining({ open: true }),
     )
-  })
-
-  it("doesn't call `renderDropDown` when not opened", () => {
-    const renderDropDown = jest.fn()
-    mount(<Dropdown renderDropDown={renderDropDown}>{jest.fn()}</Dropdown>)
-    expect(renderDropDown).not.toHaveBeenCalled()
   })
 
   it("shouldn't add a <ClickOutside /> component if `onClickOutside` is not provided", () => {
-    const renderDropDown = jest.fn(() => <p />)
+    const renderDropdown = jest.fn(() => null)
     const wrapper = mount(
-      <Dropdown open={true} renderDropDown={renderDropDown}>
-        {jest.fn()}
+      <Dropdown renderDropdown={renderDropdown}>
+        {jest.fn(() => null)}
       </Dropdown>,
     )
     expect(wrapper.find(ClickOutside)).toHaveLength(0)
   })
 
   it('should add a <ClickOutside /> component if `onClickOutside` is provided', () => {
-    const renderDropDown = jest.fn(() => <p />)
+    const renderDropdown = jest.fn(() => null)
     const wrapper = mount(
-      <Dropdown
-        onClickOutside={jest.fn()}
-        open={true}
-        renderDropDown={renderDropDown}
-      >
-        {jest.fn()}
+      <Dropdown onClickOutside={jest.fn()} renderDropdown={renderDropdown}>
+        {jest.fn(() => null)}
       </Dropdown>,
     )
     expect(wrapper.find(ClickOutside).first()).toBeDefined()