+++ packages/react-behave/src/components/Dropdown/minWidthModifier.test.js
@@ -7,6 +7,7 @@ describe('minWidthModifier', () => {
         reference: { width: 100 },
         popper: { width: 100 },
       },
+      placement: 'top',
       styles: {},
     })
     expect(res.styles.minWidth).toBe('100px')
@@ -18,6 +19,19 @@ describe('minWidthModifier', () => {
         reference: { width: 100 },
         popper: { width: 101 },
       },
+      placement: 'bottom-start',
+      styles: {},
+    })
+    expect(res.styles.minWidth).toBeUndefined()
+  })
+
+  it("doesn't add a `minWidth` if the drop down is on the left or right side of the reference", () => {
+    const res = minWidthModifier.fn({
+      offsets: {
+        reference: { width: 100 },
+        popper: { width: 99 },
+      },
+      placement: 'left',
       styles: {},
     })
     expect(res.styles.minWidth).toBeUndefined()