+++ packages/react-behave/src/components/MergeRefs/MergeRefs.js
@@ -11,15 +11,15 @@ import setRef from './setRef'
  * ## Usage
  *
  * ```jsx
- * import React, { Component } from 'react';
- * import { ComponentNeedsRef } from 'some-lib';
- * import { MergeRefs } from 'react-behave';
+ * import React, { Component } from 'react'
+ * import { ComponentNeedsRef } from 'some-lib'
+ * import { MergeRefs } from 'react-behave'
  *
  * class App extends Component {
- *   buttonRef = React.createRef();
+ *   buttonRef = React.createRef()
  *
  *   componentDidMount() {
- *     this.buttonRef.focus();
+ *     this.buttonRef.focus()
  *   }
  *
  *   render() {
@@ -31,7 +31,7 @@ import setRef from './setRef'
  *           </MergeRefs>
  *         )}
  *       </ComponentNeedsRef>
- *     );
+ *     )
  *   }
  * }
  * ```