+++ packages/react-libraries/CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Documentation
+
+- Fixed typo in README.
+
 ## 0.12.0 (July 25, 2018)
 
 ### Enhancements